:bug: moved GoogleAuth service
moved and updated GoogleAuth since it was causing issues for the api
