+++ libs/api-lib/src/lib/api-lib.module.ts
@@ -4,7 +4,6 @@ import { ApiScriptDetailController } from './controllers/script-detail/api-scrip
 import { ScriptEditorController } from './controllers/script-editor/script-editor.controller';
 import { CollectionsService } from './services/collection/collections.service';
 import { ScriptService } from './services/scripts/script.service';
-import { GoogleAuthService } from './services/GoogleAuth/google-auth.service';
 import { ScriptEditorService } from './services/editor/script-editor.service';
 import { RatingService } from './services/ratings/rating.service';
 import { CommentService } from './services/comments/comment.service';
@@ -35,7 +34,7 @@ import { MetadataSchema } from './schemas/Metadata';
     ScriptService,
     RatingService,
     CommentService,
-    GoogleAuthService
+    
   ],
   exports: [],
 })
