+++ libs/client-lib/src/lib/google-login/google-login.module.ts
@@ -6,7 +6,7 @@ import { HttpClientModule } from '@angular/common/http';
 import { OAuthModule } from 'angular-oauth2-oidc';
 
 // eslint-disable-next-line @nrwl/nx/enforce-module-boundaries
-import { GoogleAuthService } from 'libs/api-lib/src/lib/services/GoogleAuth/google-auth.service';
+import { GoogleAuthService } from 'libs/client-lib/src/lib/google-login/GoogleAuth/google-auth.service';
 
 @NgModule({
   declarations: [
