+++ libs/client-lib/src/lib/shared/shared.module.ts
@@ -4,7 +4,7 @@ import { FooterComponent } from './components/footer/footer.component';
 import { HeaderComponent } from './components/header/header.component';
 import { BggSearchService } from './services/bgg-search/bgg-search.service';
 // eslint-disable-next-line @nrwl/nx/enforce-module-boundaries
-import { GoogleAuthService } from 'libs/api-lib/src/lib/services/GoogleAuth/google-auth.service';
+import { GoogleAuthService } from 'libs/client-lib/src/lib/google-login/GoogleAuth/google-auth.service';
 import { OAuthModule } from 'angular-oauth2-oidc';
 @NgModule({
   declarations: [
